# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Persistent shared memory infrastructure for AI agents. Stores observations in PostgreSQL + pgvector, exposes MCP tools via HTTP/SSE.",
        "title": "Engram API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:37777",
    "basePath": "/",
    "paths": {
...
        "/api/events/ingest": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Receives raw tool events from Claude Code hooks. Stores in raw_events, runs deterministic Level 0 pipeline, creates an observation, and triggers async embedding.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Events"
                ],
                "summary": "Ingest tool event",
                "parameters": [
                    {
                        "description": "Tool event data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_worker.IngestRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_worker.IngestRequest": {
            "type": "object",
            "properties": {
                "project": {
                    "type": "string"
                },
                "session_id": {
                    "type": "string"
                },
                "tool_input": {},
                "tool_name": {
                    "type": "string"
                },
                "tool_result": {},
                "workstation_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-Auth-Token",
            "in": "header"
        }
    }
}